requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
boto3>=1.34,<2.0
gunicorn>=21.2,<22.0
aws-wsgi>=0.2,<0.3
